refactor: entity settings styling
