--- src/modules/CreateEntitySettings/components/PrivacyCard/PrivacyCard.styles.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components'
-
-export const FormWrapper = styled.div`
-  .form-group .form-control {
-    background-color: #e8edee;
-    -webkit-box-shadow: none;
-  }
-`
