--- src/modules/CreateEntitySettings/components/RequiredCredentialCard/RequiredCredentialCard.styles.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components'
-
-export const RemoveButton = styled.button`
-  font-weight: bold;
-  font-size: 1rem;
-  line-height: 1.2;
-  border: none;
-  background-color: transparent;
-  color: #39c3e6;
-`
