--- src/modules/CreateEntitySettings/components/StatusCard/StatusCard.styles.ts
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-export const FormWrapper = styled.div`
-  .form-group .form-control {
-    background-color: #e8edee;
-    -webkit-box-shadow: none;
-  }
-  .DateRangePicker .DateRangePickerInput__showClearDates {
-    display: flex;
-  }
-`
